update project change to ui and add to auth

+++ redux-is-takip/src/components/Tasks/Tasks.jsx
@@ -1,22 +1,37 @@
-import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { fetchEmployees } from "../../slices/employeesSlice";
 import TasksContent from "./TasksContent";
 import bgTask from "../../images/jobs.png";
 import { deleteTask, fetchTasks } from "../api/api";
+import { useEffect, useState } from "react";
+import { TaskModal } from "../modals/TaskModal";
+import { fetchEmployees } from "../../slices/employeesSlice";
 
 const Tasks = () => {
-  const tasks = useSelector((state) => state.task.tasks);
+  const { tasks } = useSelector((state) => state.task);
+  const { employees } = useSelector((state) => state.employee);
+  const [isOpen, setIsOpen] = useState(false);
   const dispatch = useDispatch();
+
   const handleTaskDelete = (id) => {
-    dispatch(deleteTask(id));
+    if (window.confirm("Emin misiniz?")) {
+      dispatch(deleteTask(id));
+    }
   };
+
+  const handleModalOpen = () => {
+    setIsOpen(true);
+  };
+  const handleModalClose = () => {
+    setIsOpen(false);
+  };
+
   useEffect(() => {
     dispatch(fetchTasks());
+    dispatch(fetchEmployees());
   }, [dispatch]);
   return (
     <div
-      className="vh-100"
+      className="h-screen"
       style={{
         background: `url(${bgTask})`,
         backgroundSize: "600px",
@@ -24,11 +39,24 @@ const Tasks = () => {
         backgroundPosition: "right bottom",
       }}
     >
-      <h4>Görevler</h4>
+      <TaskModal
+        handleModalClose={handleModalClose}
+        isOpen={isOpen}
+        employees={employees}
+      />
+      <h4 className="font-bold text-blue-400">Görevler</h4>
+      <button
+        className="text-blue-600 bg-blue-200 p-2 my-2 rounded-lg   ease-in-out hover:text-blue-200 hover:bg-blue-600 duration-500"
+        onClick={handleModalOpen}
+      >
+        Yeni Görev Ekle
+      </button>
       {tasks?.length > 0 ? (
         <TasksContent tasks={tasks} handleTaskDelete={handleTaskDelete} />
       ) : (
-        <div className="alert alert-success">Hiçbir görev bulunmamaktadır.</div>
+        <div className="text-green-700 bg-green-200 rounded-lg p-4">
+          Hiçbir görev bulunmamaktadır.
+        </div>
       )}
     </div>
   );
